agent/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
python-telegram-bot>=21.0
requests>=2.31.0
beautifulsoup4>=4.12.0
